+++ src/main/java/com/centralesupelec/chowchow/user/controllers/UserDTO.java
@@ -1,14 +1,13 @@
 package com.centralesupelec.chowchow.user.controllers;
 
 import com.centralesupelec.chowchow.likes.controllers.ShowRatingDTO;
-import com.centralesupelec.chowchow.user.domain.PremiumUserEntity;
 import com.centralesupelec.chowchow.user.domain.SubscriptionType;
 import com.centralesupelec.chowchow.user.domain.UserEntity;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.List;
 import java.util.Objects;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @JsonInclude(JsonInclude.Include.NON_NULL) // Ignore the null values when parsing into Json
@@ -16,13 +15,13 @@ public class UserDTO {
   private final Long id;
   private final String username;
   private final SubscriptionType subscriptionType;
-  private final Set<ShowRatingDTO> likedShows;
+  private final List<ShowRatingDTO> likedShows;
 
   @JsonCreator
   public UserDTO(
       @JsonProperty("id") Long id,
       @JsonProperty("username") String username,
-      @JsonProperty("likedShows") Set<ShowRatingDTO> likedShows,
+      @JsonProperty("likedShows") List<ShowRatingDTO> likedShows,
       @JsonProperty("subscriptionType") SubscriptionType subscriptionType) {
     this.id = id;
     this.username = username;
@@ -32,13 +31,8 @@ public class UserDTO {
 
   public static UserEntity toEntity(UserDTO userDTO) {
     UserEntity userEntity = new UserEntity();
-    if (userDTO.subscriptionType != null) {
-      PremiumUserEntity premiumUserEntity = new PremiumUserEntity();
-      premiumUserEntity.setSubscriptionType(userDTO.subscriptionType);
-      userEntity = premiumUserEntity;
-      userEntity.setLikedShows(null);
-    }
     userEntity.setUsername(userDTO.getUsername());
+    userEntity.setLikedShows(userDTO.likedShows.stream().map(likedShow -> U));
     return userEntity;
   }
 
@@ -52,7 +46,7 @@ public class UserDTO {
         userEntity.getUsername(),
         userEntity.getLikedShows().stream()
             .map(ShowRatingDTO::fromEntity)
-            .collect(Collectors.toSet()),
+            .collect(Collectors.toList()),
         subscriptionType);
   }
 
@@ -72,7 +66,7 @@ public class UserDTO {
     return subscriptionType;
   }
 
-  public Set<ShowRatingDTO> getLikedShows() {
+  public List<ShowRatingDTO> getLikedShows() {
     return likedShows;
   }
 }
